Remove the dependency require in service.lua
