+++ service/service.lua
@@ -465,6 +465,89 @@ function ltask.signal_handler(f)	-- root only
 	end
 end
 
+local yieldable_require; do
+	local require = _G.require
+	local loaded = package.loaded
+	local loading = {}
+	function yieldable_require(name)
+		local m = loaded[name]
+		if m ~= nil then
+			return m
+		end
+		local _, main = coroutine.running()
+		if main then
+			return require(name)
+		end
+		local filename = package.searchpath(name, package.path)
+		if not filename then
+			return require(name)
+		end
+		local modfunc = loadfile(filename)
+		if not modfunc then
+			return require(name)
+		end
+		if loading[name] then
+			error(("Another coroutine is loading `%s`."):format(name))
+		end
+		loading[name] = true
+		local ok, r = xpcall(modfunc, debug.traceback, name, filename)
+		loading[name] = nil
+		if ok then
+			if r == nil then
+				r = true
+			end
+			loaded[name] = r
+			return r
+		else
+			error(r)
+		end
+	end
+end
+
+local function init_exclusive()
+	local exclusive = require "ltask.exclusive"
+	local blocked_message
+	local retry_blocked_message
+	local function post_message(address, session, type, msg, sz)
+		if not exclusive.send(address, session, type, msg, sz) then
+			if blocked_message then
+				local n = #blocked_message
+				blocked_message[n+1] = address
+				blocked_message[n+2] = session
+				blocked_message[n+3] = type
+				blocked_message[n+4] = msg
+				blocked_message[n+5] = sz
+			else
+				blocked_message = { address, session, type, msg, sz }
+				ltask.fork(retry_blocked_message)
+			end
+		end
+	end
+	function retry_blocked_message()
+		ltask.sleep(0)
+		if not blocked_message then
+			return
+		end
+		local blocked = blocked_message
+		blocked_message = nil
+		for i = 1, #blocked, 5 do
+			local address = blocked[i]
+			local session = blocked[i+1]
+			local type    = blocked[i+2]
+			local msg     = blocked[i+3]
+			local sz      = blocked[i+4]
+			post_message(address, session, type, msg, sz)
+		end
+	end
+	function ltask.post_message(address, session, type, msg, sz)
+		post_message(address, session, type, msg, sz)
+		return true
+	end
+	function ltask.error(address, session, errobj)
+		post_message(address, session, MESSAGE_ERROR, ltask.pack(errobj))
+	end
+end
+
 function sys_service.init(t)
 	-- The first system message
 	assert(service == nil)
@@ -477,9 +560,9 @@ function sys_service.init(t)
 	local _require = _G.require
 	local filename = assert(package.searchpath(t.name, t.service_path))
 	local f = assert(loadfile(filename))
-	_G.require = require "ltask.require"
+	_G.require = yieldable_require
 	if t.exclusive then
-		require "ltask.init_exclusive"
+		init_exclusive()
 	end
 	local r = f(table.unpack(t.args))
 	_G.require = _require
